refactor: move the Annex K tables to common so codec layers stop reaching up
Issue #442's evidence list names two places where lower-level modules
depend upward on `encode`. Both have one cause: the JPEG Annex K
specification tables live in `encode::tables`, although nothing about
them is encoder-specific.

- `common::huffman_table::std_huffman_tables()` imported the four
  standard Huffman tables from `encode` — to serve the *decoder*.
- Five `simd/*` sites imported `ZIGZAG_ORDER` from `encode` for
  quantisation kernels used in both directions.

The spec data — two base quantization tables, four Huffman bits/values
pairs, and the zigzag order — moves to `common::tables`. `src/common/`
now has zero `crate::encode` references.

`encode::tables` is a public path (`pub mod encode` → `pub mod tables`),
so every moved name is re-exported from there and no caller breaks.
Encoder *policy* stays put: `quality_scale_quant_table{,_ext}` turn a base
table into a quantization table, which is a choice rather than a
specification. The three tests that validate the Annex K data moved with
the data; the quality-scaling tests stayed with the functions.

The `simd → encode` references that remain are scalar *kernel* fallbacks
for encode-direction SIMD (`encode::color::rgb_to_ycbcr_row` and friends).
That is the correct direction — a SIMD kernel falling back to its scalar
twin — not a spec-data inversion, so they are left alone.

Also files the umbrella itself as **P4-123**. #442 had no LAST_MILE entry,
which by this repository's own rule means it did not exist for the next
session reading the index. It is structured like P4-55: the workstream
detail stays in the GitHub issue, per-defect behaviour stays with the
individual P4 entries, and the index row points at both.

Measured on macOS aarch64: `cargo test --workspace --no-fail-fast` →
2472 passed, 0 failed, 1 ignored — same as before the move, as expected
for a relocation with re-exports. `cargo fmt --check` and the CI clippy
invocation are clean.

Refs #442
