+++ docs/LAST_MILE.md
@@ -112,6 +112,7 @@ The table below is the live OPEN set across all filing dates. **P4-13 / P4-14 / 
 | P4-115 | [Native 12-bit coverage claims include untested modes/layouts](last_mile/phase4.md#p4-115-native-12-bit-coverage-claims-include-untested-modes-and-sampling-layouts--open) | 4 | OPEN | Raw C matrix is six baseline layouts; S410/S24 raw-planar cases and progressive/arithmetic coverage remain missing. |
 | P4-120 | [classic-shim allocation-failure paths are unreachable from tests](last_mile/phase4.md#p4-120-classic-shim-allocation-failure-paths-are-unreachable-from-tests--open) | 4 | OPEN | `JERR_OUT_OF_MEMORY` (code 56, and its `ERREXIT1` `msg_parm` payload) is emitted by two `jpeg_mem_dest` paths that no test can force. Filed in the P4-108 review; fits P4-116's test-integrity scope. |
 | P4-121 | [lossless encode accepts a restart interval C refuses to decode](last_mile/phase4.md#p4-121-lossless-encode-accepts-a-restart-interval-c-refuses-to-decode--open) | 4 | OPEN | `jclossls.c:294-296` raises `JERR_BAD_RESTART` when `Ri % MCUs_per_row != 0`; we emit it, producing a stream `djpeg` rejects. Found while repairing the P4-116 lossless cross-check. |
+| P4-123 | [Architecture umbrella: codec plans, C-ABI state, public boundaries, SIMD dispatch](last_mile/phase4.md#p4-123-architecture-umbrella-codec-plans-c-abi-state-public-boundaries-simd-dispatch--open) | 4 | OPEN | Umbrella for the 2026-08-08 architecture programme; detail lives in GitHub #442. Five workstreams (classic C-ABI boundary, canonical encode/decode plans, public/dependency boundaries, SIMD dispatch containment, one 8/12/16-bit model). Workstream 3's spec-data inversion is done: the Annex K tables moved to `common::tables`, so `common` and `simd` no longer reach upward into `encode`. |
 | P2-G | [Downstream lab multi-version / multi-distro matrix](last_mile/backlog.md#p2-g-downstream-lab-multi-version--multi-distro-matrix--open) | Backlog | OPEN | The pinned OpenCV 4.6/Ubuntu 24.04 leg is complete; Qt5/Qt6, additional OpenCV/distro versions, immutable scheduled environments, and broader version pairs remain. |
 
 **Next up:** close the remaining P0, the create-size overwrite risk **P4-110**. Then unify errors/lifecycle through **P4-100/P4-104/P4-106** — **P4-116** closed 2026-08-08, so the compatibility evidence those results rest on is now fail-closed. After those safety/correctness blockers, close ELF versioning **P4-81**, work through **P4-84..P4-115**, and expand **[P2-G](last_mile/backlog.md#p2-g-downstream-lab-multi-version--multi-distro-matrix--open)**. P4-108 destination ownership, P4-82 restart settings, and baseline-only P4-83 smoothing are closed; progressive/arithmetic smoothing remains P4-84.
